// src/serverApi/v3/models/share-token-body-params.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ShareTokenBodyParams
 */
export interface ShareTokenBodyParams {
    /**
     * the type of the object being shared
     * @type {string}
     * @memberof ShareTokenBodyParams
     */
    parentType: ShareTokenBodyParamsParentType;
    /**
     * the id of the object being shared.
     * @type {string}
     * @memberof ShareTokenBodyParams
     */
    parentId: string;
    /**
     * when defined, the sharetoken will expire after the given number of days.
     * @type {number}
     * @memberof ShareTokenBodyParams
     */
    expiresInDays?: number | null;
    /**
     * when defined, the sharetoken will be usable exclusively by members of the users school.
     * @type {boolean}
     * @memberof ShareTokenBodyParams
     */
    schoolExclusive?: boolean | null;
}

/**
    * @export
    * @enum {string}
    */
export enum ShareTokenBodyParamsParentType {
    COURSES = 'courses',
    TASKS = 'tasks',
    LESSONS = 'lessons'
}